Refactor database
Always use None for default value (used to be 0 in add_* functions).
Set current time in SQLite, not in Python.
Update only not None values (if value is None not selecting old values from db and updating to them, but not updating this values at all).
Decided not to check types of methods' parameters (other database can do it by itself).